requirements.txt
django>=5.1
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
PyPDF2>=3.0.1
jsonschema>=4.23.0
djangorestframework-simplejwt>=5.3.0
